refactor code to use content-property-dataset
