+++ src/packages/core/content/index.ts
@@ -1,4 +1,4 @@
-export * from './workspace/index.js';
-
-export { UmbContentPropertyContext } from './content-property.context.js';
 export { UMB_CONTENT_PROPERTY_CONTEXT } from './content-property.context-token.js';
+export { UmbContentPropertyContext } from './content-property.context.js';
+export * from './property-dataset-context/content-property-dataset.context.js';
+export * from './workspace/index.js';
