+++ src/packages/documents/document-blueprints/property-dataset-context/document-blueprint-property-dataset-context.token.ts
@@ -1,13 +1,14 @@
 import { UMB_DOCUMENT_BLUEPRINT_ENTITY_TYPE } from '../entity.js';
-import type { UmbDocumentBlueprintPropertyDataContext } from './document-blueprint-property-dataset-context.js';
+import type { UmbDocumentBlueprintPropertyDatasetContext } from './document-blueprint-property-dataset-context.js';
 import type { UmbPropertyDatasetContext } from '@umbraco-cms/backoffice/property';
 import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 
 export const IsDocumentBlueprintPropertyDatasetContext = (
 	context: UmbPropertyDatasetContext,
-): context is UmbDocumentBlueprintPropertyDataContext => context.getEntityType() === UMB_DOCUMENT_BLUEPRINT_ENTITY_TYPE;
+): context is UmbDocumentBlueprintPropertyDatasetContext =>
+	context.getEntityType() === UMB_DOCUMENT_BLUEPRINT_ENTITY_TYPE;
 
 export const UMB_DOCUMENT_BLUEPRINT_PROPERTY_DATASET_CONTEXT = new UmbContextToken<
 	UmbPropertyDatasetContext,
-	UmbDocumentBlueprintPropertyDataContext
+	UmbDocumentBlueprintPropertyDatasetContext
 >('UmbVariantContext', undefined, IsDocumentBlueprintPropertyDatasetContext);
