+++ src/packages/documents/document-blueprints/workspace/document-blueprint-workspace.context.ts
@@ -1,4 +1,4 @@
-import { UmbDocumentBlueprintPropertyDataContext } from '../property-dataset-context/document-blueprint-property-dataset-context.js';
+import { UmbDocumentBlueprintPropertyDatasetContext } from '../property-dataset-context/document-blueprint-property-dataset-context.js';
 import { UMB_DOCUMENT_BLUEPRINT_ENTITY_TYPE } from '../entity.js';
 import { UmbDocumentBlueprintDetailRepository } from '../repository/index.js';
 import type {
@@ -441,8 +441,11 @@ export class UmbDocumentBlueprintWorkspaceContext
 		}
 	}
 
-	public createPropertyDatasetContext(host: UmbControllerHost, variantId: UmbVariantId) {
-		return new UmbDocumentBlueprintPropertyDataContext(host, this, variantId);
+	public createPropertyDatasetContext(
+		host: UmbControllerHost,
+		variantId: UmbVariantId,
+	): UmbDocumentBlueprintPropertyDatasetContext {
+		return new UmbDocumentBlueprintPropertyDatasetContext(host, this, variantId);
 	}
 
 	public destroy(): void {
