+++ src/packages/members/member/workspace/member-workspace.context.ts
@@ -1,6 +1,6 @@
 import { UmbMemberDetailRepository } from '../repository/index.js';
 import type { UmbMemberDetailModel, UmbMemberVariantModel, UmbMemberVariantOptionModel } from '../types.js';
-import { UmbMemberPropertyDataContext } from '../property-dataset-context/member-property-dataset-context.js';
+import { UmbMemberPropertyDatasetContext } from '../property-dataset-context/member-property-dataset-context.js';
 import { UMB_MEMBER_WORKSPACE_ALIAS } from './manifests.js';
 import { UmbMemberWorkspaceEditorElement } from './member-workspace-editor.element.js';
 import { type UmbMemberTypeDetailModel, UmbMemberTypeDetailRepository } from '@umbraco-cms/backoffice/member-type';
@@ -371,8 +371,11 @@ export class UmbMemberWorkspaceContext
 		}
 	}
 
-	public createPropertyDatasetContext(host: UmbControllerHost, variantId: UmbVariantId) {
-		return new UmbMemberPropertyDataContext(host, this, variantId);
+	public createPropertyDatasetContext(
+		host: UmbControllerHost,
+		variantId: UmbVariantId,
+	): UmbMemberPropertyDatasetContext {
+		return new UmbMemberPropertyDatasetContext(host, this, variantId);
 	}
 
 	public destroy(): void {
